Refactor: common methods and classes
