+++ src/formats/mod.rs
@@ -3,6 +3,7 @@ use std::fmt::Debug;
 use crate::pcd::PointCloudData;
 use crate::velodyne::{VelodynPoint, VelodyneBinData};
 
+pub mod bounds;
 pub mod pointxyzrgba;
 pub mod pointxyzrgbanormal;
